reusing Front and ListWithQuery in favor of custom comps. translations and minor fixes
