+++ pages/preview.js
@@ -1,6 +1,5 @@
 import React from 'react'
 import { compose } from 'react-apollo'
-import Frame from '../components/Frame'
 import Preview from '../components/Marketing/Preview'
 import withData from '../lib/apollo/withData'
 import withInNativeApp from '../lib/withInNativeApp'
@@ -17,16 +16,14 @@ const IndexPage = ({ url, t, me, isMember, inNativeIOSApp }) => {
     return <UnauthorizedPage me={me} url={url} />
   }
   const meta = {
-    pageTitle: 'Die Republik probelesen',
-    title: 'Die Republik probelesen',
-    description: 'Melden Sie sich an, um kostenlos fünf Artikel zu lesen.',
+    pageTitle: t('marketing-20/preview/page/pageTitle'),
+    title: t('marketing-20/preview/page/title'),
+    description: t('marketing-20/preview/page/description'),
     image: `${CDN_FRONTEND_BASE_URL}/static/social-media/logo.png`,
     url: `${PUBLIC_BASE_URL}/`
   }
   return (
-    <Frame raw url={url} meta={meta}>
-      <Preview url={url} />
-    </Frame>
+    <Preview url={url} meta={meta} />
   )
 }
